Updated installer

+++ install.sh
@@ -27,8 +27,7 @@ print_title() {
 authenticate() {
 	local hwid=$(get_hwid)
 
-	# ! DO NOT TOUCH ! DO NOT TOUCH ! DO NOT TOUCH !
-	local whitelist # MUST assign the variable before initializing to prevent $? not working properly
+	local whitelist
 	whitelist=$(curl -s "https://git.raptor.fun/api/whitelist?hwid=$hwid" 2>&1)
 	local status=$?
 	if [[ "$status" != 0 ]]; then
@@ -38,8 +37,7 @@ authenticate() {
 		center "\033[33mAlthough you could still install MacSploit,\033[0m"
 		center "\033[33myou will not be able to use MacSploit without a VPN.\033[0m"
 		center "Do you want to proceed? (Y/N): \c"
-		while true; do
-			read -rsn1 answer
+		while read -n 1 -s -r answer; do
 			if [[ "$answer" =~ ^[Yy]$ ]]; then
 				print_title
 				return
@@ -49,14 +47,13 @@ authenticate() {
 				exit
 			fi
 		done
-		return
 	fi
 
 	local trial=false
 	center "\033[36mWelcome to the MacSploit experience!\033[0m"
 	sleep 2
 	print_title
-	echo
+	return
 }
 
 check_requirements() {
@@ -86,7 +83,8 @@ check_requirements() {
 
 check_permissions() {
 	if [[ -d "$PWD" ]]; then
-		local error=$(rm -rf "$PWD" 2>&1)
+		local error
+		error=$(rm -rf "$PWD" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			center "\033[91mTerminal is unable to access your Temporary folder.\033[0m"
@@ -101,7 +99,8 @@ check_permissions() {
 		fi
 	fi
 
-	local error=$(mkdir "$PWD" 2>&1)
+	local error
+	error=$(mkdir "$PWD" 2>&1)
 	local status=$?
 	if echo "$error" | grep -q "Permission denied"; then
 		center "\033[91mTerminal is unable to access your Temporary folder.\033[0m"
@@ -115,7 +114,8 @@ check_permissions() {
 		exit
 	fi
 
-	local error=$(rm -f "$HOME/Downloads/ms-version.json" 2>&1)
+	local error
+	error=$(rm -f "$HOME/Downloads/ms-version.json" 2>&1)
 	local status=$?
 	if echo "$error" | grep -q "Permission denied"; then
 		if ! can_sudo; then
@@ -131,7 +131,8 @@ check_permissions() {
 		exit
 	fi
 
-	local error=$(touch "$HOME/Downloads/ms-version.json" 2>&1)
+	local error
+	error=$(touch "$HOME/Downloads/ms-version.json" 2>&1)
 	local status=$?
 	if echo "$error" | grep -q "Operation not permitted"; then
 		center "\033[91mTerminal is unable to access your Downloads folder.\033[0m"
@@ -146,7 +147,8 @@ check_permissions() {
 			echo
 		fi
 
-		local error=$(sudo touch "$HOME/Downloads/ms-version.json" 2>&1)
+		local error
+		error=$(sudo touch "$HOME/Downloads/ms-version.json" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			center "\033[91mTerminal was unable to access your Downloads folder.\033[0m"
@@ -156,6 +158,12 @@ check_permissions() {
 		elif echo "$error" | grep -q "is not in the sudoers file."; then
 			center "\033[91mTerminal was unable to access your user folder.\033[0m"
 			center "\033[91mPlease contact support for help.\033[0m"
+			echo
+			exit
+		elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+			center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+			center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+			center "\033[91mPlease try again.\033[0m"
 			echo
 			exit
 		elif [[ "$status" != 0 ]]; then
@@ -174,7 +182,8 @@ check_permissions() {
 
 	local deleted=false
 	if [[ -d "/Applications/Roblox.app" ]]; then
-		local error=$(rm -rf /Applications/Roblox.app 2>&1)
+		local error
+		error=$(rm -rf "/Applications/Roblox.app" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			if ! can_sudo; then
@@ -182,7 +191,39 @@ check_permissions() {
 				center "\033[91mPlease enter your password to grant sudo permissions.\033[0m"
 				echo
 			fi
-			sudo rm -rf /Applications/Roblox.app
+
+			local error
+			error=$(sudo rm -rf "/Applications/Roblox.app" 2>&1)
+			local status=$?
+			if echo "$error" | grep -q "Permission denied"; then
+				center "\033[91mTerminal was unable to access your Applications folder.\033[0m"
+				center "\033[91mPlease contact support for help.\033[0m"
+				echo
+				exit
+			elif echo "$error" | grep -q "is not in the sudoers file."; then
+				APPDIR="$HOME/Applications"
+
+				local error
+				error=$(mkdir "$APPDIR" 2>&1)
+				local status=$?
+				if ! echo "$error" | grep -q "File exists" && [[ "$status" != 0 ]]; then
+					center "\033[91mTerminal was unable to create an user Applications folder.\033[0m"
+					center "\033[91mThis is unexpected, please contact support.\033[0m"
+					echo
+					exit
+				fi
+			elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+				center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+				center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+				center "\033[91mPlease try again.\033[0m"
+				echo
+				exit
+			elif [[ "$status" != 0 ]]; then
+				center "\033[91mAn unknown error has occurred: C-04.\033[0m"
+				center "\033[91mThis is unexpected, please contact support.\033[0m"
+				echo
+				exit
+			fi
 		elif [[ "$status" != 0 ]]; then
 			center "\033[91mAn unknown error has occurred: C-04.\033[0m"
 			center "\033[91mThis is unexpected, please contact support.\033[0m"
@@ -194,7 +235,8 @@ check_permissions() {
 	fi
 
 	if [[ -d "$HOME/Applications/Roblox.app" ]]; then
-		local error=$(rm -rf "$HOME/Applications/Roblox.app" 2>&1)
+		local error
+		error=$(rm -rf "$HOME/Applications/Roblox.app" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			if ! can_sudo; then
@@ -202,7 +244,39 @@ check_permissions() {
 				center "\033[91mPlease enter your password to grant sudo permissions.\033[0m"
 				echo
 			fi
-			sudo rm -rf "$HOME/Applications/Roblox.app"
+
+			local error
+			error=$(sudo rm -rf "$HOME/Applications/Roblox.app" 2>&1)
+			local status=$?
+			if echo "$error" | grep -q "Permission denied"; then
+				center "\033[91mTerminal was unable to access your user folder.\033[0m"
+				center "\033[91mPlease contact support for help.\033[0m"
+				echo
+				exit
+			elif echo "$error" | grep -q "is not in the sudoers file."; then
+				APPDIR="$HOME/Applications"
+
+				local error
+				error=$(mkdir "$APPDIR" 2>&1)
+				local status=$?
+				if ! echo "$error" | grep -q "File exists" && [[ "$status" != 0 ]]; then
+					center "\033[91mTerminal was unable to create an user Applications folder.\033[0m"
+					center "\033[91mThis is unexpected, please contact support.\033[0m"
+					echo
+					exit
+				fi
+			elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+				center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+				center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+				center "\033[91mPlease try again.\033[0m"
+				echo
+				exit
+			elif [[ "$status" != 0 ]]; then
+				center "\033[91mAn unknown error has occurred: C-05.\033[0m"
+				center "\033[91mThis is unexpected, please contact support.\033[0m"
+				echo
+				exit
+			fi
 		elif [[ "$status" != 0 ]]; then
 			center "\033[91mAn unknown error has occurred: C-05.\033[0m"
 			center "\033[91mThis is unexpected, please contact support.\033[0m"
@@ -214,7 +288,8 @@ check_permissions() {
 	fi
 
 	if [[ -d "/Applications/MacSploit.app" ]]; then
-		local error=$(rm -rf "/Applications/MacSploit.app" 2>&1)
+		local error
+		error=$(rm -rf "/Applications/MacSploit.app" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			if ! can_sudo; then
@@ -222,7 +297,39 @@ check_permissions() {
 				center "\033[91mPlease enter your password to grant sudo permissions.\033[0m"
 				echo
 			fi
-			sudo rm -rf "/Applications/MacSploit.app"
+
+			local error
+			error=$(sudo rm -rf "/Applications/MacSploit.app" 2>&1)
+			local status=$?
+			if echo "$error" | grep -q "Permission denied"; then
+				center "\033[91mTerminal was unable to access your Applications folder.\033[0m"
+				center "\033[91mPlease contact support for help.\033[0m"
+				echo
+				exit
+			elif echo "$error" | grep -q "is not in the sudoers file."; then
+				APPDIR="$HOME/Applications"
+
+				local error
+				error=$(mkdir "$APPDIR" 2>&1)
+				local status=$?
+				if ! echo "$error" | grep -q "File exists" && [[ "$status" != 0 ]]; then
+					center "\033[91mTerminal was unable to create an user Applications folder.\033[0m"
+					center "\033[91mThis is unexpected, please contact support.\033[0m"
+					echo
+					exit
+				fi
+			elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+				center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+				center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+				center "\033[91mPlease try again.\033[0m"
+				echo
+				exit
+			elif [[ "$status" != 0 ]]; then
+				center "\033[91mAn unknown error has occurred: C-06.\033[0m"
+				center "\033[91mThis is unexpected, please contact support.\033[0m"
+				echo
+				exit
+			fi
 		elif [[ "$status" != 0 ]]; then
 			center "\033[91mAn unknown error has occurred: C-06.\033[0m"
 			center "\033[91mThis is unexpected, please contact support.\033[0m"
@@ -234,7 +341,8 @@ check_permissions() {
 	fi
 
 	if [[ -d "$HOME/Applications/MacSploit.app" ]]; then
-		local error=$(rm -rf "$HOME/Applications/MacSploit.app" 2>&1)
+		local error
+		error=$(rm -rf "$HOME/Applications/MacSploit.app" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			if ! can_sudo; then
@@ -242,7 +350,39 @@ check_permissions() {
 				center "\033[91mPlease enter your password to grant sudo permissions.\033[0m"
 				echo
 			fi
-			sudo rm -rf "$HOME/Applications/MacSploit.app"
+
+			local error
+			error=$(sudo rm -rf "$HOME/Applications/MacSploit.app" 2>&1)
+			local status=$?
+			if echo "$error" | grep -q "Permission denied"; then
+				center "\033[91mTerminal was unable to access your Applications folder.\033[0m"
+				center "\033[91mPlease contact support for help.\033[0m"
+				echo
+				exit
+			elif echo "$error" | grep -q "is not in the sudoers file."; then
+				APPDIR="$HOME/Applications"
+
+				local error
+				error=$(mkdir "$APPDIR" 2>&1)
+				local status=$?
+				if ! echo "$error" | grep -q "File exists" && [[ "$status" != 0 ]]; then
+					center "\033[91mTerminal was unable to create an user Applications folder.\033[0m"
+					center "\033[91mThis is unexpected, please contact support.\033[0m"
+					echo
+					exit
+				fi
+			elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+				center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+				center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+				center "\033[91mPlease try again.\033[0m"
+				echo
+				exit
+			elif [[ "$status" != 0 ]]; then
+				center "\033[91mAn unknown error has occurred: C-07.\033[0m"
+				center "\033[91mThis is unexpected, please contact support.\033[0m"
+				echo
+				exit
+			fi
 		elif [[ "$status" != 0 ]]; then
 			center "\033[91mAn unknown error has occurred: C-07.\033[0m"
 			center "\033[91mThis is unexpected, please contact support.\033[0m"
@@ -262,8 +402,7 @@ check_permissions() {
 		center "\033[91mThis is required since MacSploit runs on top of Rosetta.\033[0m"
 		echo
 		center "Do you want to install Rosetta? (Y/N): \c"
-		while true; do
-			read -rsn1 answer
+		while read -n 1 -s -r answer; do
 			if [[ "$answer" =~ ^[Yy]$ ]]; then
 				echo
 				echo
@@ -306,8 +445,7 @@ check_version() {
 		center "\033[33mThanks to update hooks, MacSploit may still function for a few days.\033[0m"
 		echo
 		center "Do you want to proceed? (Y/N): \c"
-		while true; do
-			read -rsn1 answer
+		while read -n 1 -s -r answer; do
 			if [[ "$answer" =~ ^[Yy]$ ]]; then
 				break
 			elif [[ "$answer" =~ ^[Nn]$ ]]; then
@@ -337,7 +475,8 @@ install_roblox() {
 	rm -rf "$PWD/RobloxPlayer.app/Contents/MacOS/Roblox.app"
 	rm -rf "$PWD/RobloxPlayer.app/Contents/MacOS/RobloxPlayerInstaller.app"
 
-	local error=$(mv "$PWD/RobloxPlayer.app" "$APPDIR/Roblox.app" 2>&1)
+	local error
+	error=$(mv "$PWD/RobloxPlayer.app" "$APPDIR/Roblox.app" 2>&1)
 	local status=$?
 	if echo "$error" | grep -q "Permission denied"; then
 		if ! can_sudo; then
@@ -346,7 +485,8 @@ install_roblox() {
 			echo
 		fi
 
-		local error=$(sudo mv "$PWD/RobloxPlayer.app" "$APPDIR/Roblox.app" 2>&1)
+		local error
+		error=$(sudo mv "$PWD/RobloxPlayer.app" "$APPDIR/Roblox.app" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			center "\033[91mTerminal was unable to access your Applications folder.\033[0m"
@@ -356,7 +496,18 @@ install_roblox() {
 		elif echo "$error" | grep -q "is not in the sudoers file."; then
 			APPDIR="$HOME/Applications"
 
-			local error=$(mv "$PWD/RobloxPlayer.app" "$APPDIR/Roblox.app" 2>&1)
+			local error
+			error=$(mkdir "$APPDIR" 2>&1)
+			local status=$?
+			if ! echo "$error" | grep -q "File exists" && [[ "$status" != 0 ]]; then
+				center "\033[91mTerminal was unable to create an user Applications folder.\033[0m"
+				center "\033[91mThis is unexpected, please contact support.\033[0m"
+				echo
+				exit
+			fi
+
+			local error
+			error=$(mv "$PWD/RobloxPlayer.app" "$APPDIR/Roblox.app" 2>&1)
 			local status=$?
 			if echo "$error" | grep -q "Permission denied"; then
 				center "\033[91mTerminal was unable to access your user folder.\033[0m"
@@ -369,6 +520,12 @@ install_roblox() {
 				echo
 				exit
 			fi
+		elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+			center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+			center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+			center "\033[91mPlease try again.\033[0m"
+			echo
+			exit
 		elif [[ "$status" != 0 ]]; then
 			center "\033[91mAn unknown error has occurred: C-08.\033[0m"
 			center "\033[91mThis is unexpected, please contact support.\033[0m"
@@ -434,6 +591,7 @@ patch_roblox() {
 		echo
 		exit
 	fi
+
 	local output=$("$PWD/insert_dylib" "$APPDIR/Roblox.app/Contents/MacOS/interpose.dylib" "$APPDIR/Roblox.app/Contents/MacOS/RobloxPlayer" "$APPDIR/Roblox.app/Contents/MacOS/RobloxPlayer" --overwrite --strip-codesig --all-yes)
 	if ! echo "$output" | grep -q "Added LC_LOAD_DYLIB to"; then
 		center "\033[91mTerminal was unable to patch RobloxPlayer.\033[0m"
@@ -487,7 +645,8 @@ clean_up() {
 	center "🧹 \033[1;36mCleaning Up...\033[0m"
 	echo
 
-	local error=$(echo "$VERSION_INFO" >"$HOME/Downloads/ms-version.json" 2>&1)
+	local error
+	error=$(echo "$VERSION_INFO" >"$HOME/Downloads/ms-version.json" 2>&1)
 	local status=$?
 	if echo "$error" | grep -q "Permission denied"; then
 		if ! can_sudo; then
@@ -496,7 +655,8 @@ clean_up() {
 			echo
 		fi
 
-		local error=$(sudo echo "$VERSION_INFO" >"$HOME/Downloads/ms-version.json" 2>&1)
+		local error
+		error=$(sudo echo "$VERSION_INFO" >"$HOME/Downloads/ms-version.json" 2>&1)
 		local status=$?
 		if echo "$error" | grep -q "Permission denied"; then
 			center "\033[91mTerminal was unable to access your Downloads folder.\033[0m"
@@ -506,6 +666,12 @@ clean_up() {
 		elif echo "$error" | grep -q "is not in the sudoers file."; then
 			center "\033[91mTerminal was unable to access your user folder.\033[0m"
 			center "\033[91mPlease contact support for help.\033[0m"
+			echo
+			exit
+		elif [[ "$error" == "sudo: 3 incorrect password attempts" ]]; then
+			center "\033[91mYou have entered an incorrect password 3 times.\033[0m"
+			center "\033[91mYour password looks invisible but is still being typed.\033[0m"
+			center "\033[91mPlease try again.\033[0m"
 			echo
 			exit
 		elif [[ "$status" != 0 ]]; then
